pyproject.toml
[project]
name = "erpnext-mcp-server"
version = "0.1.5"
description = "An MCP server for interacting with an ERPNext instance."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "markdownify",
    "readabilipy"
]

[project.scripts]
erpnext-mcp-server = "mcp.run"
